perf(retrieval): simplify BM25 IDF mathematical formula and refactor tests
